+++ src/main/java/org/loverde/geographiccoordinate/GeographicCoordinateException.java
@@ -29,20 +29,11 @@ public class GeographicCoordinateException extends Exception {
                                  LATITUDE_MINUTES_AND_SECONDS_MUST_BE_ZERO = LATITUDE + MINUTES_AND_SECONDS_MUST_BE_ZERO + MAX_LATITUDE_DEGREES,
                                  LATITUDE_MINUTES_RANGE = LATITUDE + MINUTES_RANGE,
                                  LATITUDE_SECONDS_RANGE = LATITUDE + SECONDS_RANGE,
-                                 LATITUDE_1_NULL = LATITUDE + " 1 is null",
-                                 LATITUDE_2_NULL = LATITUDE + " 2 is null",
 
                                  LONGITUDE_DEGREES_RANGE = LONGITUDE + DEGREES_RANGE + MAX_LONGITUDE_DEGREES,
                                  LONGITUDE_MINUTES_AND_SECONDS_MUST_BE_ZERO = LONGITUDE + MINUTES_AND_SECONDS_MUST_BE_ZERO + MAX_LONGITUDE_DEGREES,
                                  LONGITUDE_MINUTES_RANGE = LONGITUDE + MINUTES_RANGE,
-                                 LONGITUDE_SECONDS_RANGE = LONGITUDE + SECONDS_RANGE,
-                                 LONGITUDE_1_NULL = LONGITUDE + " 1 is null",
-                                 LONGITUDE_2_NULL = LONGITUDE + " 2 is null",
-
-                                 POINT_1_NULL = "Point 1 is null",
-                                 POINT_2_NULL = "Point 2 is null",
-
-                                 UNIT_NULL = "Unit is null";
+                                 LONGITUDE_SECONDS_RANGE = LONGITUDE + SECONDS_RANGE;
    }
 
    public GeographicCoordinateException() {